separate fetch implementation
