+++ src/system.js
@@ -83,88 +83,7 @@ function applyPaths(loader, name) {
   return outPath;
 }
 
-(function() {
-  var fetchTextFromURL;
-  if (typeof XMLHttpRequest != 'undefined') {
-    fetchTextFromURL = function(url, fulfill, reject) {
-      var xhr = new XMLHttpRequest();
-      var sameDomain = true;
-      var doTimeout = false;
-      if (!('withCredentials' in xhr)) {
-        // check if same domain
-        var domainCheck = /^(\w+:)?\/\/([^\/]+)/.exec(url);
-        if (domainCheck) {
-          sameDomain = domainCheck[2] === window.location.host;
-          if (domainCheck[1])
-            sameDomain &= domainCheck[1] === window.location.protocol;
-        }
-      }
-      if (!sameDomain && typeof XDomainRequest != 'undefined') {
-        xhr = new XDomainRequest();
-        xhr.onload = load;
-        xhr.onerror = error;
-        xhr.ontimeout = error;
-        xhr.onprogress = function() {};
-        xhr.timeout = 0;
-        doTimeout = true;
-      }
-      function load() {
-        fulfill(xhr.responseText);
-      }
-      function error() {
-        reject(xhr.statusText + ': ' + url || 'XHR error');
-      }
-
-      xhr.onreadystatechange = function () {
-        if (xhr.readyState === 4) {
-          if (xhr.status === 200 || (xhr.status == 0 && xhr.responseText)) {
-            load();
-          } else {
-            error();
-          }
-        }
-      };
-      xhr.open("GET", url, true);
-
-      if (doTimeout)
-        setTimeout(function() {
-          xhr.send();
-        }, 0);
-
-      xhr.send(null);
-    };
-  }
-  else if (typeof require != 'undefined') {
-    var fs;
-    fetchTextFromURL = function(url, fulfill, reject) {
-      if (url.substr(0, 8) != 'file:///')
-        throw 'Only file URLs of the form file:/// allowed running in Node.';
-      fs = fs || require('fs');
-      if (isWindows)
-        url = url.replace(/\//g, '\\').substr(8);
-      else
-        url = url.substr(7);
-      return fs.readFile(url, function(err, data) {
-        if (err)
-          return reject(err);
-        else
-          fulfill(data + '');
-      });
-    };
-  }
-  else {
-    throw new TypeError('No environment fetch API available.');
-  }
-
-  // inline Object.create-style class extension
-  function LoaderProto() {}
-  LoaderProto.prototype = Loader.prototype;
-  SystemLoader.prototype = new LoaderProto();
-
-  SystemLoader.prototype.fetch = function(load) {
-    return new Promise(function(resolve, reject) {
-      fetchTextFromURL(load.address, resolve, reject);
-    });
-  };
-
-})();
+// inline Object.create-style class extension
+function LoaderProto() {}
+LoaderProto.prototype = Loader.prototype;
+SystemLoader.prototype = new LoaderProto();
